backend/pyproject.toml
[project]
name = "agent-hub"
version = "0.1.0"
description = "一个基于 LangChain 和 LangGraph 的多 Agent 协作系统"
requires-python = ">=3.10"
dependencies = [
    # LangChain 核心
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",

    # LangGraph - 多 Agent 编排
    "langgraph>=0.2.0",

    # 向量数据库
    "chromadb>=0.5.0",

    # LLM 提供商 - 阿里百炼 (兼容 OpenAI 接口)
    "langchain-openai>=0.3.0",

    # 文档加载
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",

    # 工具
    "python-dotenv>=1.0.0",
    "tiktoken>=0.7.0",

    # Web 服务
    "flask>=3.0.0",
    "flask-cors>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend/src/core", "backend/src/agents", "backend/src/knowledge", "backend/src/session"]
